[Authentification] Continuing securing actions

+++ grails-app/controllers/outofbounds/AnswerController.groovy
@@ -102,11 +102,11 @@ class AnswerController {
         def answer = Answer.findById(params.answer_id)
         def question = answer.question
 
-        if (answer.canUserDeletePost(getAuthenticatedUser())) {
+        if (answer && answer.canUserDeletePost(getAuthenticatedUser())) {
             answerService.delete(answer)
-            flash.message = message(code: 'answer.delete_success')
+            flash.message = message(code: 'post.delete_success', args: ["answer"])
         } else {
-            flash.message = message(code: 'answer.delete_not_authorized')
+            flash.message = message(code: 'post.delete_not_authorized', args: ["answer"])
         }
         redirect(uri: "/question/show?question_id=${question.id}")
     }
